+++ ui/src/app/components/runtime-configuration-panel/create-panel.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 
 import { Profile, RuntimeStatus, Workspace } from 'generated/fetch';
 
-import { Button } from 'app/components//buttons';
 import { UIAppType } from 'app/components/apps-panel/utils';
+import { Button } from 'app/components/buttons';
 import { EnvironmentInformedActionPanel } from 'app/components/common-env-conf-panels/environment-informed-action-panel';
 import { styles } from 'app/components/common-env-conf-panels/styles';
 import { FlexRow } from 'app/components/flex';
